:tractor: relocate files for coherence
