+++ moban/core/context.py
@@ -1,7 +1,8 @@
 import os
 import copy
 
-from moban import utils, reporter, constants, exceptions
+from moban import utils, constants, exceptions
+from moban.externals import reporter
 from moban.program_options import OPTIONS
 from moban.data_loaders.manager import merge, load_data
