+++ moban/mobanfile/__init__.py
@@ -4,9 +4,10 @@ import sys
 import logging
 from collections import OrderedDict
 
-from moban import core, reporter, constants
+from moban import core, constants
 from lml.utils import do_import
 from moban.utils import verify_the_existence_of_directories
+from moban.externals import reporter
 from moban.deprecated import handle_copy, handle_requires
 from moban.mobanfile.targets import (
     parse_targets,
