+++ moban/mobanfile/targets.py
@@ -1,7 +1,8 @@
 import uuid
 import logging
 
-from moban import core, reporter, constants, exceptions
+from moban import core, constants, exceptions
+from moban.externals import reporter
 from moban.definitions import TemplateTarget
 from moban.mobanfile.templates import handle_template
