+++ tests/test_copy_engine.py
@@ -1,9 +1,9 @@
 import os
 
 import fs.path
-from moban import file_system
 from moban.copy import ContentForwardEngine
 from nose.tools import eq_
+from moban.externals import file_system
 
 
 class TestContentForwardEngine:
